prompt-generation/requirements.txt
pandas==2.0.3
matplotlib==3.9.2
numpy==1.26.4
groq==0.12.0
python-dotenv==1.0.1
openai==1.50.2